+++ src/lib.rs
@@ -125,6 +125,7 @@ pub use ticker::{ClientHandle, Ticker, TickerBuilder};
 pub use domains::CryptoCoin;
 #[cfg(any(
     feature = "alphavantage",
+    feature = "bls",
     feature = "fiscaldata",
     feature = "fred",
     feature = "polygon",
@@ -244,6 +245,7 @@ pub use models::commodities::CommodityQuote;
 pub use models::crypto::CryptoQuote;
 #[cfg(any(
     feature = "alphavantage",
+    feature = "bls",
     feature = "fiscaldata",
     feature = "fred",
     feature = "polygon",